+++ app/components/ProductSearchResult.jsx
@@ -1,40 +1,12 @@
-"use client";
-// import Link from "next/link";
+
+import Link from "next/link";
 import Image from "next/image";
-import { useEffect, useRef } from "react";
-import { useRouter } from "next/navigation";
 
-const ProductSearchResult = ({ searchProduct, setSearchProduct, setSearch }) => {
 
-    const searchResultRef = useRef(null);
-    const router = useRouter();
-     // Function to handle click outside the modal
-     useEffect(() => {
-        const handleClickOutside = (event) => {
-            if (
-                searchResultRef.current &&
-                !searchResultRef.current.contains(event.target)
-            ) {
-                setSearchProduct([]);
-                setSearch("");
-            }
-        };
-        if (typeof window !== "undefined") {
-            document.addEventListener("mousedown", handleClickOutside);
-        }
-        return () => {
-            document.removeEventListener("mousedown", handleClickOutside);
-        };
-    }, [setSearchProduct, setSearch]);
-
-    const handleProductClick = (product) => {
-        setSearchProduct([]);
-        setSearch("");
-        router.push(`/products/get-product-details?outlet_id=${product?.outlet_id}&product_id=${product?.product_id}`);
-    };
+const ProductSearchResult = ({ searchProduct }) => {
 
     return (
-        <div ref={searchResultRef} className="product-search-modal-area">
+        <div className="product-search-modal-area">
             <div className="product-search-modal-content">
                 <div className="search-modal-title">
                     <h4>products</h4>
@@ -45,10 +17,9 @@ const ProductSearchResult = ({ searchProduct, setSearchProduct, setSearch }) => 
                             searchProduct.map((product, index) => (
                                 <li
                                     key={`${product?.product_name}-${product?.slug}-${index}`}
-                                    onClick={() => handleProductClick(product)}
-                                    style={{ cursor:"pointer" }}
+                                    
                                 >
-                                    {/* <Link href={`/products/get-product-details?outlet_id=${product?.outlet_id}&product_id=${product?.product_id}`}> */}
+                                    <Link href={`/products/get-product-details?outlet_id=${product?.outlet_id}&product_id=${product?.product_id}`}>
                                         <div className="search-modal-info-inner d-flex align-content-center gap-4">
                                             <div className="search-modal-info-img">
                                                 <Image
@@ -64,7 +35,7 @@ const ProductSearchResult = ({ searchProduct, setSearchProduct, setSearch }) => 
                                                 </strong>
                                             </div>
                                         </div>
-                                    {/* </Link> */}
+                                    </Link>
                                 </li>
                             ))}
                     </ul>
